moved view item dialog to its own class
